encoding: Generate CommandCreator enum and boilerplate using derive
