+++ lorawan-encoding/src/maccommandcreator.rs
@@ -618,93 +618,8 @@ pub fn build_mac_commands<T: AsMut<[u8]>>(
     Ok(i)
 }
 
-macro_rules! mac_cmds_creator_enum {
-    (
-        $outer_vis:vis enum $outer_type:ident$(<$outer_lifetime:lifetime>),* {
-        $(
-            $name:ident, $creator:ident
-        )*
-    }
-    ) => {
-        #[allow(dead_code, missing_docs)]
-        #[derive(Debug)]
-        #[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
-        $outer_vis enum $outer_type$(<$outer_lifetime>)* {
-            $(
-                $name($creator),
-            )*
-        }
-        #[allow(clippy::len_without_is_empty)]
-        impl$(<$outer_lifetime>)* $outer_type$(<$outer_lifetime>)* {
-            /// Get the length.
-            pub fn len(&self) -> usize {
-                match self {
-                    $(
-                        Self::$name(creator) => creator.len(),
-                    )*
-                }
-            }
-            /// Build.
-            pub fn build(&self) -> &[u8] {
-                match *self {
-                    $(
-                        Self::$name(ref v) => v.build(),
-                    )*
-                }
-            }
-        }
-        impl SerializableMacCommand for $outer_type$(<$outer_lifetime>)* {
-            fn payload_bytes(&self) -> &[u8] {
-                &self.build()[1..]
-            }
+#[doc(inline)]
+pub use crate::maccommands::DownlinkMacCommandCreator;
 
-            fn cid(&self) -> u8 {
-                match self {
-                    $(
-                        Self::$name(creator) => creator.cid(),
-                    )*
-                }
-            }
-
-            fn payload_len(&self) -> usize {
-                self.len() - 1
-            }
-        }
-    }
-}
-
-mac_cmds_creator_enum! {
-    pub enum UplinkMacCommandCreator {
-        // 1.0.0
-        LinkCheckReq, LinkCheckReqCreator
-        LinkADRAns, LinkADRAnsCreator
-        DutyCycleAns, DutyCycleAnsCreator
-        RXParamSetupAns, RXParamSetupAnsCreator
-        DevStatusAns, DevStatusAnsCreator
-        NewChannelAns, NewChannelAnsCreator
-        RXTimingSetupAns, RXTimingSetupAnsCreator
-        // 1.0.2
-        TXParamSetupAns, TXParamSetupAnsCreator
-        DlChannelAns, DlChannelAnsCreator
-        // 1.0.3
-        DeviceTimeReq, DeviceTimeReqCreator
-    }
-}
-
-mac_cmds_creator_enum! {
-    pub enum DownlinkMacCommandCreator {
-        // 1.0.0
-        LinkCheckAns, LinkCheckAnsCreator
-        LinkADRReq, LinkADRReqCreator
-        DutyCycleReq, DutyCycleReqCreator
-        RXParamSetupReq, RXParamSetupReqCreator
-        DevStatusReq, DevStatusReqCreator
-        NewChannelReq, NewChannelReqCreator
-        RXTimingSetupReq, RXTimingSetupReqCreator
-        // 1.0.2
-        TXParamSetupReq, TXParamSetupReqCreator
-        DlChannelReq, DlChannelReqCreator
-        // 1.0.3
-        DeviceTimeAns, DeviceTimeAnsCreator
-    }
-}
+#[doc(inline)]
+pub use crate::maccommands::UplinkMacCommandCreator;
